+++ jest.config.js
@@ -21,7 +21,7 @@ const customJestConfig = {
 
   moduleNameMapper: {
     "@/(.*)$": "<rootDir>/src/$1",
-    "@ui/(.*)$": "<rootDir>/components/ui/$1",
+    "@ui/(.*)$": "<rootDir>/src/components/ui/$1",
   },
   testEnvironment: "jest-environment-jsdom",
 };
